Split the folder wizard sources
